requirements.txt
datefinder-lexpredict==0.6.2
Pillow==7.1.1
pytesseract==0.3.3
python-dateutil==2.8.1
pytz==2019.3
regex==2017.9.23
six==1.14.0
